requirements.txt
spotipy
python-dotenv
winotify
requests